Daf's "select-dataset" implementation
Keep in mind that this commit will not provide the whole functionality,
this is only the implementation and not the full feature's integration.

> The integration includes an update about the nteract's redux store
too. By updating is meant the process of merge Daf's epics and reducers
with nteract's ones. Which is not what we are going to aim in this
changes, that process will comes with the "daf-nteract-packages/ducks"
package implementation.

- Adds "select-dataset" package to "daf-nteract-packages".

- Adds "DatasetPopover" as "DafMenu" component in cell's toolbar found
in "notebook-app-component/src/toolbar.tsx".

- Adds custom style rules to ".bp3-menu" class in "DatasetPopover"
component.

- Implements "DatasetListSuggest" components.

- Implements "datasetList" and "selectDataset" ducks.