Daf's "login" feature implementation update
Since we are trying to implement some private Daf's features we have to
get a user authentication token in order to get it done.

- Closes issue #15

- Resolves issue #16

- Fixes bug #20

- Re-enables "fetchContent" and "fetchKernelspecs" dispatches in order
to unlock the access of nteract's features, so you are not forced
anymore to login with a Daf's user account in order to use notebooks.
Only if you want to use Daf's features you have login.

- Updates "@daf-nteract-packages/login" to 0.2.0

- Updates "@daf-nteract-packages/ducks" to 0.3.0

- Adds "loggedUserDataSelector", "usernameSelector", "tokenSelector",
"isUserLogged", "rejectLoginEpic", "removeLocals" to
"daf-nteract-packages/login" duck.

- Updates "requestLoginEpic", "logoutEpic", "rejectLogin", " in
"daf-nteract-packages/login" duck.

- Re-implements "loginEpic" in "daf-nteract-packages/login" duck.

- Removes "initializeLogin", "LOGIN_INIT" in "daf-nteract-packages/login" duck.

- Renames "LogoutButton" component to "LoginLogoutButton".

- Re-implements "LoginDialog" opening.

- Adds "IDatasetPopoverProps" type in
"daf-nteract-packages/select-dataset" types

- Updates "DatasetPopover" rendering logic

- Renames "DatasetPopover" component to "DatasetPopoverContainer"